Bugfix: make search tests more reliable
This fixes a test which could intermitently fail.

The cause of the failures is down to using Faker to generate the descriptions of the test investigations, and then search query. On ocassion, by chance the search query could match more than one result (in part due to the use of the OR operator in the search query), which caused the test to fail.

This fix makes that scenario a lot less likely by explicitly setting descriptions on each of the investigations, and explicitly setting a search query which should only ever match one of those.