Bugfix: sets yesterday’s date when today is the first of the month (#237)
Previously when the day was the first of the month, clicking "Yesterday" set the date as the 0th of the month. This fix correctly sets it to the last day of the previous month, by using the [`setDate()`](https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Date/setDate) function on the date object, which if you pass `0`, sets the date object to the last day of the previous month (and previous year, if necessary).