Make .create_and_send_invite tests more explicit
Different developers had different approaches on this.

I'm happy to move all the shared test setup into the "before hook"
as the context is compact enough to show what is happening where.
Lets leave "subject" (or whatever named subject we use) for class
instances and just use "let" when needed.

At the end of the day they're the same and at least we can trust
"subject" calls to fall under its default "instance of the described
class" behaviour.