Improve mobile number validation on registration
We want to be flexible but not allow random stuff introduced as mobile
numbers.
Due to that we're just validating that the mobile number contains, at
least, 10 digits.

In case of, even passing this validation, the provided number is
incorrect, Notify service won't accept it and the user will not receive
their code.
In that case they will be able to change their number in a further
step.