Redirect already activated users to root page
Some users that already completed/activated their account after
receiving the invitation, will bookmark the invitation URL as their
way to access PSD service.
Given that, we want to rediret users that are either already activated
or even signed in to the service main page when they try to access the
service through the invite link.