Creating an account from an invitation (#294)
Provides the implementation for the user account registration completion after following the link in the invitation email.

When landing into the complete registration page a visitor can:

1. Be a non signed in visitor with a valid invitation token for an user that haven't previously completed its registration.
In this case we provide the user with the form to fill its name, mobile phone and password and complete its registration into the service.

2. Be a signed user with a different user from the invited one, following the link a valid token for the invited user.
This situation can happen, for example, when multiple users share a computer (eg: Invited user trying to access from the same computer as the user who invited her/him uses).
In this case we notify the user that either can either be redirected to the home page as its current user or log out from the current user and get to complete the registration completion form.  

3. Be already signed in as the invited user, or not being signed in but the user corresponding to the invitation has been already activated.
This may happen when people bookmark the invitation link and use that link to access the service.
In that case we redirect them to the home page. They can sign in from there if needed to, but they won't be able to re-complete the registration page.

4. Access to the correct link, but having an expired invitation token.
If the invitation was sent more than 14 days before the user visit the complete registration page, the user won't be able to complete the registration and he will get told to ask for a new invitation or contact OPSS support.

5. Access with an incorrect invitation token.
In this case the user will get an error page and won't be able to complete the registration.

About validations
Some basic mobile and password validations have been introduced, there is a room for improvement here, but since we will require 2 Factor Authentication users will need to introduce a valid UK mobile phone if they want to be able to receive the authentication code and use the service.

Co-authored-by: Frankie Roberto <frankie@frankieroberto.com>