Check whether account has been setup, not activated (#319)
The "activated" flag should only be set after the user has accepted the declaration, so we need a different check within the account setup flow to check that the user hasn't already set up a password and phone number. For this, I've added a new method on the user called `has_completed_registration?` which simply checks whether the user has a password, mobile number and name.