Bugfix: don't check whether user has viewed introduction on help pages (#314)
Users should be able to view the "terms and conditions" and "privacy notice" pages at any point, regardless of whether they have accepted the declaration or not, are signed in or not, or have viewed the introduction slides or not.

This is particularly important as the declaration page links to the terms and conditions and privacy notice pages, which users are agreeing to by accepting the declaration. However currently non-OPSS users are unable to view these at this point, as the pages are redirecting to the introduction slides, which then redirect back to the declaration page.