Use Env var to Enable/Disable 2FA
In order to get more flexibility about when 2FA is required or not
(2FA may be an annoyance on test/development).

Defaulted the TEST environment to not use 2FA, and only use it for
specific feature tests that assert over the sign in process itself.

2FA was clashing with Request specs, as the stubs for sign_up weren't
covering the 2FA bit and the redirections after signing up were pointing
to 2FA page.
Inverting to "only 2FA when specifically required" approach avoids this.