Enforce antivirus checked document on Feature test
The "Delete attachment" feature test was having failures when running
the full Rspec suite.
When executed individually the test was passing.

Seems the issue was that the attached document was not being marked
as analyzed by the antivirus and checked as "safe", so the link
for edition or removing the attachment wasn't displayed in the
document page.

I wasn't able to figure out why exactly this was happening, so decided
to enforce having a document analyzed/approved by the antivirus in the
factory for this test, solving the issue.