Handle requesting a password reset for an invited email
If a password reset is requested for an e-mail address which has been invited, but hasn't yet set up an account, then we want to send a different e-mail from the regular reset password one, depending on the invitation status.

If the invitation is still valid (within the set time period), then we should re-send the invitation e-mail, allowing the user to continue with setting up their account.

If the invitation has expired, then we should block the password reset process, and instead send a different e-mail asking the user to ask a colleague to re-send their invitation.