Default Notify Api Key env var to empty
The env var from Cloud Foundry user provided service instances is not
available on buildpack stage.
This causes the deploys to fail when compiling assets on deploy.

I understand the downside on defaulting this API KEY:
The application won't blow up on startup when the value is not set,
risking forgetting to set the env var before pushing this into different
environments.
But is that or not being able to deploy with the current
infrastructure.