Move all captions outside of h1 elements
This looks the same visually, but means that when the page title is "read out", eg by screen reader software using the headings for navigation or document outlines, then the page title will be less verbose.

This is espcieally significant where the `<h1>` is also embedded within a form `<legend>` as that can be read out multiple times.