Include "Error: " prefix in page title when there are validation errors
Adding an "Error: " prefix to the `<title>` tag on pages with validation errors means that screen readers announce the error as soon as possible.

To achive this, all pages have switched to using the `page_title` helper, with the `errors` parameter set to `true` if there are any validation errors present in forms included on the pages.