Include error prefix in page titles (#396)
Adding an "Error: " prefix to the `<title>` tag on pages with validation errors, so that screen readers announce the error as soon as possible. See [GOV.UK Design System pattern on validation](https://design-system.service.gov.uk/patterns/validation/).

To implement this, I've switched all views to use a new `page_title` helper, which accepts an optional `error` boolean that adds the prefix if truthy. The helper also includes the `content_for` call to DRY the views.