Remove unnecessary referal check
When an user is able to reach this point, it will already passed the
reset_password_token check. So will come from the reset password
email containing this token in the url.

This will avoid users accessing this password edit page without having
gone through the reset password process and receiving the email.

Given this, checking the referral to ensure signed user comes from
2FA becames unnecessary.