Allow user to sign out and reset other user password
Given than userA requests the password reset and uses a computer shared
with userB (that is already signed in) to open the "reset password"
link, we end with a valid link containing the correct reset password
token, but with the wrong user requesting it.

Instead of showing "invalid link", we are going to reproduce the
behaviour for completing registration for user invitation link.

This behaviour consists in offering the already signed in user to
either continue as itself (being redirected to the homepage) or to sign
out to proceed as userA and complete the password change.