Avoid sending 2FA code when not requesting 2FA
We only want to send sms with 2FA code for users that are going to be
redirected to 2FA page in order to fil the form with the received code.

Our sign in business logic dictates some cases where we don't want
the sign in submission to reach 2FA, like when the user didn't verified
its phone during the registration from an invitation email link.

The ruby gem used for 2FA adds a hook to successful Warden
authentication calls, so whenever a Warden authentication happens, a
2FA code gets send to the user (if the user is marked as requiring 2FA).

Devise helpers like "current_user" and "user_signed_in?" internally
do a "warden.authenticate", triggering the 2FA code being sent.

Our application controller has multiple before filters calling those
Devise methods. So, when submitting a Sign in form, the filters where
being triggered and the 2FA code was being sent even before we could
run our business logic to dictate if an user gets directed to 2FA page
or not.

This was causing users not being directed to 2FA page and being shown
a login error (as the business logic dictates) but still getting an
SMS with a 2FA code for each login submission.

The solution consists in suppressing those calls to Devise methods
based in existing user until the point where we actually want to sign
in an user. In that point we set up those variables/invoke those calls.

This way the 2FA code will be sent only when we explicitely authenticate
the user, instead of being an uncontrolled side effect.