Refactor session creation method
The method in charge of handling user sign-in submissions has become
complex with multiple branching for dealing with edge cases and
different ways to redirect after specific failures.

As a result of all this logic, I found the method had to follow and
change.

As an attempt to improve its maintainability I:

- Re-shaped it so the different edge cases and errors provoke an early
  return one by one, while defaulting at the end to the happy path with
  a successful sign in. This is done to allow the reader to follow a
  linear process of the checks that would cause a sign-in failure.

- Documented all these checks in the method declaration, so future
  maintainers can understand better how our Sign-in processs work.

- Extracted a couple private methods encapsulating checks/logic to
  show the intention of the code bit through the naming of the method.

- Added comments where considered necessary.