Bugfix: Avoid sending 2FA code when not requesting 2FA (#458)
* Avoid sending 2FA code when not requesting 2FA

We only want to send sms with 2FA code for users that are going to be
redirected to 2FA page in order to fill the form with the received code.

Our sign in business logic dictates some cases where we don't want
the sign in submission to reach 2FA, like when the user didn't verified
its phone during the registration from an invitation email link.

The ruby gem used for 2FA adds a hook to successful Warden
authentication calls, so whenever a Warden authentication happens, a
2FA code gets send to the user (if the user is marked as requiring 2FA).

Devise helpers like "current_user" and "user_signed_in?" internally
do a "warden.authenticate", triggering the 2FA code being sent.

Our application controller has multiple before filters calling those
Devise methods. So, when submitting a Sign in form, the filters where
being triggered and the 2FA code was being sent even before we could
run our business logic to dictate if an user gets directed to 2FA page
or not.

This was causing users not being directed to 2FA page and being shown
a login error (as the business logic dictates) but still getting an
SMS with a 2FA code for each login submission.

The solution consists in suppressing those calls to Devise methods
based in existing user until the point where we actually want to sign
in an user. In that point we set up those variables/invoke those calls.

This way the 2FA code will be sent only when we explicitely authenticate
the user, instead of being an uncontrolled side effect.

* Refactor session creation method

The method in charge of handling user sign-in submissions has become
complex with multiple branching for dealing with edge cases and
different ways to redirect after specific failures.

As a result of all this logic, I found the method hard to follow and
change.

As an attempt to improve its maintainability we:

- Re-shaped it so the different edge cases and errors provoke an early
  return one by one, while defaulting at the end to the happy path with
  a successful sign in. This is done to allow the reader to follow a
  linear process of the checks that would cause a sign-in failure.

- Extracted a couple private methods encapsulating checks/logic to
  show the intention of the code bit through the naming of the method.

- Added comments where considered necessary.

Co-Authored-By: Frankie Roberto <frankie@frankieroberto.com>

* Extract authentication success/failure to handlers

Extract the logic for handling authentication success and failure to
their own separated methods.