Migrate Enquiry System test to Feature test
The System test was written in Minitest and was missing the new changes
asking for Coronavirus.

The team decided to not write more Minitest tests and use Rspec instead.
As we already had a Feature test for Enquiries creation, I extended it
to cover some of the bits that were being tested in the System test
and missing in the feature one.

Then removed the system tests for Enquiry.