Allow to mark users as "deleted"
We want to introduce the possibility of marking users as "deleted".
The user register will still exist in DB but this flag will, at some
point, effectively invalidate the user through the application.

At this point we just provide a rake task that will allow to mark
an used as deleted by providing users email or id.

As DB wasn't enforcing unique user emails, added the corresponding
uniqueness index restriction in DB.