Declute Interactor from logic belonging to Task
Things like searching the user by id or email are more bounded to the
actual command rake task than to the business logic associated to
deleting an user.

So, by extracting all the rake related work to the task and leaving
the interactor with less responsibilities, we achieve a simpler
service, more prone to be independently used by other callers besides
the rake task. Plus we simplified the tests quite a bit.