Group Devise overriden methods by Modules
In this way we can undertand better where these methods were originally
belonging and what are they related with.

Also this is a step towards possible future refactors like extract
groups of overriden methods belonging to the same Devise module to a
separate class inheriting from Devise one so the User model can use
it as concern without needing to have the overrides inside the User
class file.