Share Github actions gem cache between workflows
Bundle install is going to generate the samei output for Rspec and
Minitest workflows.

There is no point to keep different cached versions of the gems for
Rspec, System and Minitest workflows when, on a Gemfile update event,
the first workflow to run will generate a gemfile cache valid for the
other two.

This commits uses a single gem cache between the 3 of them. The key
will be the same and only become invalidated when the Gemfile.lock
file changes.