Bugfix: default "Assigned to me team" to 'unchecked' if param missing
This fixes the bug whereby if you click the "Your cases" link on the homepage, you also see cases assigned to your team.

This was caused because the filter builds the query by assuming that the team ID of your team should be added to the query unless the `assigned_to_team_0` param has the value of `unchecked`. But because the link from the homepage doesn't set this param (it just sets `/cases?assigned_to_me=checked`), the param had the value of nil instead of `unchecked`.

A private `set_default_assignee_filter` sets the params values to `unchecked` if they are missing, but the `assigned_to_team_0` was missing. Adding this fixes the bug.