keep original operation when resending 2FA code
The resend code controller was falling under the default operation for
2FA check purposes.
That was causing issues when the original operation was having a shorter
expiry time than the default.

Eg:
- Invite user has x hours expiry.
- Sign in has x days expiry.
- Sign in is the default operation.
- When enquiried for 2fa on inviting an user action, we request a new
  code to be sent to us.
- Nothing happens, because when checking again under the default
  operation (with the 2FA code being valid for x days), the user has a
  "recent enough" 2FA and does not need to get authenticated again.

To solve this, we will use the last operation associated to the user,
that will be updated when last required for 2FA.
So the "resend code" will carry this operation, and the rules associated
to it will be respected.