Allow to change unverified phones on 2FA
When users that have not completed the registration and verified their
phones try to go though 2FA for registration completion, they may have
issuesd like having introduced a wrong phone and never receiving the
sms with the code needed to pass 2FA.

To avoid that lock situation we will allow users that still didn't
verified their phone to fill a new phone number when requesting
their 2FA code to be re-send to them.

The code will be sent to this new phone and the phone number will
be updated for the user.