Extract domains whitelisting OTPs to a yaml file
Extract the regular expressions defining the domains where secondary
authentication One Time Password codes can be whitelisted to a constants
YAML file.

We do not want to set this in ENV VARS per environment due to strict
security requirements. To avoid an easy (and hard to trace) mess up
because an env var was incorrectly set.
We neither want urls hardcoded in our business logic.
Extracting these regular expressions defining our urls into a YAML
file of constants is a solution that sits in the middle of the
mentioned two approaches.