Only allow collaborators to update case (#606)
This restricts most of the "update" actions on a case to users belonging to teams added to the case. The exception is updating the case status (eg open / close), whether it’s restricted or not, and who the owners and collaborators are, which are all restricted to users of the case owner team.