Fixes a breaking test caused by a change in date format. (#670)
We switched from using `%d %B %Y`  (eg `01 June 2020`) to `'%e %B %Y` (eg ` 1 June 2020`) but forgot to update the tests. This only became apparent on the 1st of the month. :smile:

Had to add an `lstrip` to the tests to strip the leading space, as `%e` uses space-padding, but Capybara collapses multiple whitespaces together.