New flow for adding supporting information
The "Add supporting information" link from the case supporting
information tab will allow to add correspondence, generic attachments,
images, test results, corrective actions and comments over the case.

The "add supporting information" and "add correspondence" pages
act as a middlemand allowing the user to select the specific type of
information to be added to the case and they will just redirect the
user to the apropriate page (the pages for each specific action are
already created).