Restrict visibility of generic attachments
Only users belonging to a team that has permissions on the investigation
will be able to view the details of the investigation generic
attachments when visiting the investigation supporting information tab.

Other users will be aware of the attachment existence but won't be able
to view the preview, title or description from them.