Adding views for correspondence (#687)
This adds three new views for viewing the details of "correspondence" added to the case. These are linked to from the Activity timeline, but in future will be linked to from the Supporting Information tab.

Unlike test results and corrective actions, these views are only available to users whose teams have been added to the case, as the correspondence may contain personal contact details.