Rename controllers
This renames some controllers, to better handle the conflict between the controllers that handle creating things (eg corrective actions) and those that handle viewing the things. Conventionally these would be the same controller, but because the creation controllers use the `show` action as part of the creation flow (via WickedWizard), these conflict.

This PR follows a suggestion from @slorek to rename the creation controllers as `RecordThingController` so that the view controllers can be `ThingController`, instead of `ViewThingController`. This will allow us to more easily merge the creation controllers back into the regular controllers, when we get around to refactoring away from using WickedWizard.