Set specific back links for edit/delete attachment
When editing or deleting an attachment, the "back" link was pointing
to the overview page of the business/product/case.
Instead we want to point it to:
- Attachments tab of the product/business.
- Supporting information tab of the case when the attachment is a
  generic document.
- Images tab of the case when the attachment is an image.

This way the navigation becomes more natural instead of forcing us to
take a further click in order to go back to the correct tab.