Document upload back link directs to attachments
Instead to sending the user back to the overview of the
investigation/product/business, now the back link in the document
upload page will send the user to the attachments or supporting
information tab depending on if the document upload is over a
product/business/investigation.

Sadly at this point we won't have an attached document to check the
type, so we won't be able to direct the user to the images tab
when accessing the upload document page from there.

Images attachments page is planned to be separated from the generic
attachments page, so this will be resolved in the future.