Normalise custom risk levels against standard ones
When an user sets a custom risk level it will be compared against the
list of standard ones in order to just set/use the matching standard one
if possible.

The match is done ignoring spaces and case differences. So if the
standard level is "Low risk" and the custom level is "low    risk  ",
the recorded level will be "Low risk".