Bugfix: highlight investigation actions radios if there's an error
This fixes a bug whereby the radios weren’t getting the error class added if the form was submitted with no option selected.

This happened because a) the form didn't have a model associated with it, and b) the key of the model and the fieldset didn't match (`investigation_action` vs `action`). The latter was renamed because `action` is a reserved keyword within Rails params (denoting `POST` vs `PATCH`).