Make risk level form easier to understand
Or at least is what I pretend to.
The idea is that the form level contains the business logic that
transforms the risk level selection or custom level into a single
field in DB and also reads the value from DB and produces the
attributes that will be displayed in the form.