Exclude "other" risk value when needed
We do not want "other" to ve a valid risk level value when we register
the Audit activity title, as in that case we want to default to the
custom risk level instead.

For presenting the radio buttons with the options, Other will be
excluded from the default radios generation, as we want to manually
add it with specific behaviour.