Add prompt to update case risk level (if different from risk assessment) (#825)
This adds an additional question, when adding a risk assessment to a case which contains a different risk level from the existing case risk level, asking whether or not the user wants to set the case risk level to match the risk assessment.

Note that this happens after the risk assessment has already been saved to the case, and so if the user doesn’t answer, then the risk assessment is still saved. If the user does choose to update the case risk level, then an additional activity entry is added to the case, and an additional email notification will be sent to the case owner.