Convert record email correspondence flow to 1 page form (#907)
This converts the "email correspondence" flow from being a 2-page flow with a confirmation screen, to a 1-page form with no confirmation screen. This is to make it more consistent with the other flows (eg risk assessments), and to enable an editing feature to be implemented more easily (ie without sending users back through a multi-page flow).