Add a new validated barcode field, and rename the existing one (#954)
This change renames the existing "Barcode or serial number" field as "Other product identifiers", in recognition of the reality that this field currently contains a mixture of barcodes, serial numbers, model numbers, ASINs, eBay item numbers, and other unidentified codes.

It also adds a new "Barcode number" field which accepts a single barcode number (of type UPC, EAN or GTIN). This number is validated to make sure that it is of the right format (usually 12 or 13 digits) and has a valid check digit.  If valid, it is stored in a standardised 13 digit format (12 digit barcodes are padded with a leading 0).

Multiple barcodes for a single product are not currently supported, as this requires further work to understand how best to support product variants.

The barcodes are not checked for uniqueness, as we do not yet allow a single product to be associated with multiple cases.

A separate rake task moves any existing valid barcodes within the "Barcode or serial number" field to the new barcode field.