Add product brand as a new field (#966)
This adds a new field to capture the 'brand' of a product, so that this can be captured separately from the product name.

This is in anticipation of the brand becoming a separate lookup table, so that a single brand can be associated with many products. This may help us to better detect duplicate products in future, as well as making the data easier to analyse.