PRISM: add risk assessment to case
Adds functionality to add a risk assessment to a case via the case page as well as creating a new case to which the risk assessment is automatically added.

Also improves functionality when adding a risk assessment to a case via the risk assessment dashboard.

Adds a new mandatory assessment name field to help identify risk assessments and changes existing lists to use it.

Displays PRISM risk assessment added to a case on the supporting information page for that case.