Return only unique investigation IDs for case export
The includes(:... for the initial eager load of records was duplicating the id for each, due to the inner joins. The pluck(:id) then just runs a select on the whole collection, which returns ID diplicates for each case. Adding a .distinct here then removes these duplicates at the query layer